Finishing mobile app

+++ backend/src/database/migrations/00_create_points.ts
@@ -8,10 +8,11 @@ export async function up(knex: Knex){
     table.string('name').notNullable();
     table.string('email').notNullable();
     table.string('whatsapp').notNullable();
-    table.decimal('latitude').notNullable();
-    table.decimal('longitude').notNullable();
+    table.float('latitude').notNullable();
+    table.float('longitude').notNullable();
     table.string('city').notNullable();
     table.string('uf', 2).notNullable();
+
   
   })
 }
